Draws lines

+++ Assets/Scripts/FitToScreen.cs
@@ -7,7 +7,7 @@ public class FitToScreen : MonoBehaviour {
 		SpriteRenderer spriteRenderer = gameObject.GetComponent<SpriteRenderer>();
 		float spriteWidth = spriteRenderer.bounds.size.x;
 		float screenHeight = Camera.main.orthographicSize * 2;
-		float screenWidth = Screen.width / Screen.height * screenHeight;
+		float screenWidth = (float)Screen.width / (float)Screen.height * screenHeight;
 		float ratio = spriteWidth / screenWidth;
 		transform.localScale = Vector3.one * ratio;
 	}
